Refactor App.tsx into components and a custom hook
- Extracted page header into a new `PageHeader` component.
- Extracted page footer and action buttons into a new `PageFooter` component.
- Extracted all question management logic (state, data fetching, and actions) into a `useQuestionManagement` custom hook.
- Simplified `App.tsx` to be a container component that composes the new hook and components.
